Refactored bisection search into Numerical Methods file
